requirements.txt
pandas
streamlit
streamlit_chat
streamlit_extras
langchain
langchain_community
langchain-core
python-dotenv
huggingface_hub
crewai
torch
crewai_tools
transformers
tools
pulp
numpy
prophet
matplotlib